Deduplicate matrixInput/fractionInput/mathInput shared logic
Extract the logic shared between the internal math-input "cell" components
backing <matrixInput> (_matrixComponentInput) and <fractionInput>
(_fractionInputComponent) into a new utils/mathComponentInput.js:

- returnMathComponentInputConfigStateVariableDefinitions (parsing/
  number-display/width config read from the parent, plus componentType
  and focused)
- returnMathComponentInputDisplayStateVariableDefinitions (valueForDisplay,
  text, rawRendererValue)
- returnInputValueChangedStateVariableDefinitions (valueChanged/
  immediateValueChanged)
- shared updateValue / focusChanged actions
- defineSubmitAnswerExternalAction (used by all three parent inputs)
- returnMathInputParsingAttributes (matrixInput + fractionInput)

Each leaf keeps only what genuinely differs (where value/immediateValue
come from, and its accessible shortDescription). Behavior is unchanged;
the only schema delta is a shared generic `format` attribute description.
