+++ packages/doenetml-worker-javascript/src/components/FractionInput.js
@@ -1,6 +1,6 @@
 import Input from "./abstract/Input";
 import me from "math-expressions";
-import { deepCompare, convertValueToMathExpression } from "@doenet/utils";
+import { convertValueToMathExpression } from "@doenet/utils";
 import BaseComponent from "./abstract/BaseComponent";
 import {
     buildNumberDisplayParameters,
@@ -8,12 +8,16 @@ import {
     returnNumberDisplayAttributes,
     returnNumberDisplayStateVariableDefinitions,
 } from "../utils/numberDisplay";
+import { roundForDisplay } from "../utils/math";
 import {
-    latexToMathFactory,
-    normalizeLatexString,
-    roundForDisplay,
-    stripLatex,
-} from "../utils/math";
+    defineSubmitAnswerExternalAction,
+    mathComponentInputFocusChanged,
+    mathComponentInputUpdateValue,
+    returnInputValueChangedStateVariableDefinitions,
+    returnMathComponentInputConfigStateVariableDefinitions,
+    returnMathComponentInputDisplayStateVariableDefinitions,
+    returnMathInputParsingAttributes,
+} from "../utils/mathComponentInput";
 
 const blankMath = () => me.fromAst("\uff3f");
 
@@ -21,23 +25,7 @@ export class FractionInput extends Input {
     constructor(args) {
         super(args);
 
-        this.externalActions = {};
-
-        //Complex because the stateValues isn't defined until later
-        Object.defineProperty(this.externalActions, "submitAnswer", {
-            enumerable: true,
-            get: async function () {
-                let answerAncestor = await this.stateValues.answerAncestor;
-                if (answerAncestor !== null) {
-                    return {
-                        componentIdx: answerAncestor.componentIdx,
-                        actionName: "submitAnswer",
-                    };
-                } else {
-                    return;
-                }
-            }.bind(this),
-        });
+        defineSubmitAnswerExternalAction(this);
     }
 
     static componentType = "fractionInput";
@@ -81,47 +69,7 @@ export class FractionInput extends Input {
                 "parseScientificNotation",
             ],
         };
-        attributes.format = {
-            description: "Input format for the numerator and denominator.",
-            createComponentOfType: "text",
-            createStateVariable: "format",
-            defaultValue: "text",
-            public: true,
-            toLowerCase: true,
-            validValues: [
-                {
-                    value: "text",
-                    description: "Plain-text math notation (e.g., `x^2 + 1`).",
-                },
-                {
-                    value: "latex",
-                    description: "LaTeX-formatted math (e.g., `x^{2} + 1`).",
-                },
-            ],
-        };
-        attributes.functionSymbols = {
-            description: "Symbols treated as function names when parsing.",
-            createComponentOfType: "textList",
-            createStateVariable: "functionSymbols",
-            defaultValue: ["f", "g"],
-            public: true,
-        };
-        attributes.splitSymbols = {
-            description:
-                "Whether multi-character symbols are split into a product of variables.",
-            createComponentOfType: "boolean",
-            createStateVariable: "splitSymbols",
-            defaultValue: true,
-            public: true,
-        };
-        attributes.parseScientificNotation = {
-            description:
-                "Whether to parse expressions like 1e3 as scientific notation.",
-            createComponentOfType: "boolean",
-            createStateVariable: "parseScientificNotation",
-            defaultValue: false,
-            public: true,
-        };
+        Object.assign(attributes, returnMathInputParsingAttributes());
 
         Object.assign(attributes, returnNumberDisplayAttributes());
 
@@ -586,8 +534,8 @@ export default class FractionComponentInput extends BaseComponent {
 
         Object.assign(this.actions, {
             updateRawValue: this.updateRawValue.bind(this),
-            updateValue: this.updateValue.bind(this),
-            focusChanged: this.focusChanged.bind(this),
+            updateValue: mathComponentInputUpdateValue.bind(this),
+            focusChanged: mathComponentInputFocusChanged.bind(this),
         });
     }
 
@@ -612,212 +560,14 @@ export default class FractionComponentInput extends BaseComponent {
     static returnStateVariableDefinitions() {
         let stateVariableDefinitions = super.returnStateVariableDefinitions();
 
-        stateVariableDefinitions.minWidth = {
-            forRenderer: true,
-            returnDependencies: () => ({
-                fractionInputAncestor: {
-                    dependencyType: "ancestor",
-                    componentType: "fractionInput",
-                    variableNames: ["minComponentWidth"],
-                },
+        // configuration read from the parent fractionInput (parsing,
+        // number-display, width), plus componentType and focused
+        Object.assign(
+            stateVariableDefinitions,
+            returnMathComponentInputConfigStateVariableDefinitions({
+                parentComponentType: "fractionInput",
             }),
-            definition({ dependencyValues }) {
-                if (dependencyValues.fractionInputAncestor) {
-                    return {
-                        setValue: {
-                            minWidth:
-                                dependencyValues.fractionInputAncestor
-                                    .stateValues.minComponentWidth,
-                        },
-                    };
-                } else {
-                    return { setValue: { minWidth: 0 } };
-                }
-            },
-        };
-
-        // don't specify attributes on fractionComponentInput
-        // instead gets these state variables from the parent fractionInput:
-        // format, functionSymbols, splitSymbols, parseScientificNotation
-        // displayDigits, displayDecimals, displaySmallAsZero, unionFromU
-        stateVariableDefinitions.format = {
-            returnDependencies: () => ({
-                parentFormat: {
-                    dependencyType: "parentStateVariable",
-                    parentComponentType: "fractionInput",
-                    variableName: "format",
-                },
-            }),
-            definition({ dependencyValues }) {
-                return { setValue: { format: dependencyValues.parentFormat } };
-            },
-        };
-
-        stateVariableDefinitions.functionSymbols = {
-            returnDependencies: () => ({
-                parentFunctionSymbols: {
-                    dependencyType: "parentStateVariable",
-                    parentComponentType: "fractionInput",
-                    variableName: "functionSymbols",
-                },
-            }),
-            definition({ dependencyValues }) {
-                return {
-                    setValue: {
-                        functionSymbols: dependencyValues.parentFunctionSymbols,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.splitSymbols = {
-            returnDependencies: () => ({
-                parentSplitSymbols: {
-                    dependencyType: "parentStateVariable",
-                    parentComponentType: "fractionInput",
-                    variableName: "splitSymbols",
-                },
-            }),
-            definition({ dependencyValues }) {
-                return {
-                    setValue: {
-                        splitSymbols: dependencyValues.parentSplitSymbols,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.parseScientificNotation = {
-            returnDependencies: () => ({
-                parentParseScientificNotation: {
-                    dependencyType: "parentStateVariable",
-                    parentComponentType: "fractionInput",
-                    variableName: "parseScientificNotation",
-                },
-            }),
-            definition({ dependencyValues }) {
-                return {
-                    setValue: {
-                        parseScientificNotation:
-                            dependencyValues.parentParseScientificNotation,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.displayDigits = {
-            returnDependencies: () => ({
-                parentDisplayDigits: {
-                    dependencyType: "parentStateVariable",
-                    parentComponentType: "fractionInput",
-                    variableName: "displayDigits",
-                },
-            }),
-            definition({ dependencyValues, usedDefault }) {
-                let result = {
-                    setValue: {
-                        displayDigits: dependencyValues.parentDisplayDigits,
-                    },
-                };
-
-                if (usedDefault.parentDisplayDigits) {
-                    result.markAsUsedDefault = { displayDigits: true };
-                }
-
-                return result;
-            },
-        };
-
-        stateVariableDefinitions.displayDecimals = {
-            returnDependencies: () => ({
-                parentDisplayDecimals: {
-                    dependencyType: "parentStateVariable",
-                    parentComponentType: "fractionInput",
-                    variableName: "displayDecimals",
-                },
-            }),
-            definition({ dependencyValues, usedDefault }) {
-                let result = {
-                    setValue: {
-                        displayDecimals: dependencyValues.parentDisplayDecimals,
-                    },
-                };
-
-                if (usedDefault.parentDisplayDecimals) {
-                    result.markAsUsedDefault = { displayDecimals: true };
-                }
-
-                return result;
-            },
-        };
-
-        stateVariableDefinitions.displaySmallAsZero = {
-            returnDependencies: () => ({
-                parentDisplaySmallAsZero: {
-                    dependencyType: "parentStateVariable",
-                    parentComponentType: "fractionInput",
-                    variableName: "displaySmallAsZero",
-                },
-            }),
-            definition({ dependencyValues }) {
-                return {
-                    setValue: {
-                        displaySmallAsZero:
-                            dependencyValues.parentDisplaySmallAsZero,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.padZeros = {
-            returnDependencies: () => ({
-                parentPadZeros: {
-                    dependencyType: "parentStateVariable",
-                    parentComponentType: "fractionInput",
-                    variableName: "padZeros",
-                },
-            }),
-            definition({ dependencyValues }) {
-                return {
-                    setValue: {
-                        padZeros: dependencyValues.parentPadZeros,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.avoidScientificNotation = {
-            returnDependencies: () => ({
-                parentAvoidScientificNotation: {
-                    dependencyType: "parentStateVariable",
-                    parentComponentType: "fractionInput",
-                    variableName: "avoidScientificNotation",
-                },
-            }),
-            definition({ dependencyValues }) {
-                return {
-                    setValue: {
-                        avoidScientificNotation:
-                            dependencyValues.parentAvoidScientificNotation,
-                    },
-                };
-            },
-        };
-
-        stateVariableDefinitions.unionFromU = {
-            returnDependencies: () => ({
-                parentUnionFromU: {
-                    dependencyType: "parentStateVariable",
-                    parentComponentType: "fractionInput",
-                    variableName: "unionFromU",
-                },
-            }),
-            definition({ dependencyValues }) {
-                return {
-                    setValue: { unionFromU: dependencyValues.parentUnionFromU },
-                };
-            },
-        };
+        );
 
         // get prefill from parent fractionInput, depending on part
         stateVariableDefinitions.prefill = {
@@ -841,59 +591,10 @@ export default class FractionComponentInput extends BaseComponent {
             },
         };
 
-        stateVariableDefinitions.valueChanged = {
-            public: true,
-            hasEssential: true,
-            defaultValue: false,
-            shadowingInstructions: {
-                createComponentOfType: "boolean",
-            },
-            returnDependencies: () => ({}),
-            definition() {
-                return { useEssentialOrDefaultValue: { valueChanged: true } };
-            },
-            inverseDefinition({ desiredStateVariableValues }) {
-                return {
-                    success: true,
-                    instructions: [
-                        {
-                            setEssentialValue: "valueChanged",
-                            value: Boolean(
-                                desiredStateVariableValues.valueChanged,
-                            ),
-                        },
-                    ],
-                };
-            },
-        };
-
-        stateVariableDefinitions.immediateValueChanged = {
-            public: true,
-            hasEssential: true,
-            defaultValue: false,
-            shadowingInstructions: {
-                createComponentOfType: "boolean",
-            },
-            returnDependencies: () => ({}),
-            definition() {
-                return {
-                    useEssentialOrDefaultValue: { immediateValueChanged: true },
-                };
-            },
-            inverseDefinition({ desiredStateVariableValues }) {
-                return {
-                    success: true,
-                    instructions: [
-                        {
-                            setEssentialValue: "immediateValueChanged",
-                            value: Boolean(
-                                desiredStateVariableValues.immediateValueChanged,
-                            ),
-                        },
-                    ],
-                };
-            },
-        };
+        Object.assign(
+            stateVariableDefinitions,
+            returnInputValueChangedStateVariableDefinitions(),
+        );
 
         stateVariableDefinitions.value = {
             hasEssential: true,
@@ -1020,288 +721,11 @@ export default class FractionComponentInput extends BaseComponent {
             },
         };
 
-        stateVariableDefinitions.valueForDisplay = {
-            returnDependencies: () => ({
-                value: {
-                    dependencyType: "stateVariable",
-                    variableName: "value",
-                },
-                displayDigits: {
-                    dependencyType: "stateVariable",
-                    variableName: "displayDigits",
-                },
-                displayDecimals: {
-                    dependencyType: "stateVariable",
-                    variableName: "displayDecimals",
-                },
-                displaySmallAsZero: {
-                    dependencyType: "stateVariable",
-                    variableName: "displaySmallAsZero",
-                },
-            }),
-            definition: function ({ dependencyValues }) {
-                let rounded = roundForDisplay({
-                    value: dependencyValues.value,
-                    dependencyValues,
-                });
-
-                return {
-                    setValue: { valueForDisplay: rounded },
-                };
-            },
-        };
-
-        stateVariableDefinitions.text = {
-            public: true,
-            description: "The current value as a text string.",
-            shadowingInstructions: {
-                createComponentOfType: "text",
-            },
-            returnDependencies: () => ({
-                valueForDisplay: {
-                    dependencyType: "stateVariable",
-                    variableName: "valueForDisplay",
-                },
-                padZeros: {
-                    dependencyType: "stateVariable",
-                    variableName: "padZeros",
-                },
-                displayDigits: {
-                    dependencyType: "stateVariable",
-                    variableName: "displayDigits",
-                },
-                displayDecimals: {
-                    dependencyType: "stateVariable",
-                    variableName: "displayDecimals",
-                },
-                avoidScientificNotation: {
-                    dependencyType: "stateVariable",
-                    variableName: "avoidScientificNotation",
-                },
-            }),
-            definition: function ({ dependencyValues }) {
-                let params = buildNumberDisplayParameters({
-                    padZeros: dependencyValues.padZeros,
-                    displayDigits: dependencyValues.displayDigits,
-                    displayDecimals: dependencyValues.displayDecimals,
-                    avoidScientificNotation:
-                        dependencyValues.avoidScientificNotation,
-                });
-                return {
-                    setValue: {
-                        text: dependencyValues.valueForDisplay.toString(params),
-                    },
-                };
-            },
-        };
-
-        // raw value from renderer
-        stateVariableDefinitions.rawRendererValue = {
-            description: "The raw value used by the renderer.",
-            forRenderer: true,
-            hasEssential: true,
-            shadowVariable: true,
-            defaultValue: "",
-            provideEssentialValuesInDefinition: true,
-            public: true,
-            shadowingInstructions: {
-                createComponentOfType: "text",
-            },
-            additionalStateVariablesDefined: [
-                {
-                    variableName: "lastValueForDisplay",
-                    hasEssential: true,
-                    shadowVariable: true,
-                    defaultValue: null,
-                    set: convertValueToMathExpression,
-                },
-            ],
-            returnDependencies: () => ({
-                // include immediateValue for inverse definition
-                immediateValue: {
-                    dependencyType: "stateVariable",
-                    variableName: "immediateValue",
-                },
-                valueForDisplay: {
-                    dependencyType: "stateVariable",
-                    variableName: "valueForDisplay",
-                },
-            }),
-            definition({ dependencyValues, essentialValues }) {
-                // use deepCompare of trees rather than equalsViaSyntax
-                // so even tiny numerical differences that within double precision are detected
-                if (
-                    essentialValues.rawRendererValue === undefined ||
-                    !deepCompare(
-                        essentialValues.lastValueForDisplay.tree,
-                        dependencyValues.valueForDisplay.tree,
-                    )
-                ) {
-                    let rawRendererValue = stripLatex(
-                        dependencyValues.valueForDisplay.toLatex(),
-                    );
-                    if (rawRendererValue === "\uff3f") {
-                        rawRendererValue = "";
-                    }
-                    return {
-                        setValue: {
-                            rawRendererValue,
-                            lastValueForDisplay:
-                                dependencyValues.valueForDisplay,
-                        },
-                        setEssentialValue: {
-                            rawRendererValue,
-                            lastValueForDisplay:
-                                dependencyValues.valueForDisplay,
-                        },
-                    };
-                } else {
-                    return {
-                        useEssentialOrDefaultValue: {
-                            rawRendererValue: true,
-                            lastValueForDisplay: true,
-                        },
-                    };
-                }
-            },
-            async inverseDefinition({
-                desiredStateVariableValues,
-                stateValues,
-                essentialValues,
-            }) {
-                const calculateMathExpressionFromLatex = async (text) => {
-                    let expression;
-
-                    text = normalizeLatexString(text, {
-                        unionFromU: await stateValues.unionFromU,
-                    });
-
-                    // replace ^25 with ^{2}5, since mathQuill uses standard latex conventions
-                    // unlike math-expression's latex parser
-                    text = text.replace(/\^(\w)/g, "^{$1}");
-
-                    let fromLatex = latexToMathFactory({
-                        functionSymbols: await stateValues.functionSymbols,
-                        splitSymbols: await stateValues.splitSymbols,
-                        parseScientificNotation:
-                            await stateValues.parseScientificNotation,
-                    });
-
-                    try {
-                        expression = fromLatex(text);
-                    } catch (e) {
-                        expression = me.fromAst("\uFF3F");
-                    }
-                    return expression;
-                };
-
-                let instructions = [];
-
-                if (
-                    typeof desiredStateVariableValues.rawRendererValue ===
-                    "string"
-                ) {
-                    let currentValue = essentialValues.rawRendererValue;
-                    let desiredValue =
-                        desiredStateVariableValues.rawRendererValue;
-
-                    if (currentValue !== desiredValue) {
-                        instructions.push({
-                            setEssentialValue: "rawRendererValue",
-                            value: desiredValue,
-                        });
-                    }
-
-                    let currentMath =
-                        await calculateMathExpressionFromLatex(currentValue);
-                    let desiredMath =
-                        await calculateMathExpressionFromLatex(desiredValue);
-
-                    // use deepCompare of trees rather than equalsViaSyntax
-                    // so even tiny numerical differences that within double precision are detected
-                    if (!deepCompare(desiredMath.tree, currentMath.tree)) {
-                        instructions.push({
-                            setDependency: "immediateValue",
-                            desiredValue: desiredMath,
-                            treatAsInitialChange: true, // so does not change value
-                        });
-                    }
-                } else {
-                    // since desired value was not a string, it must be a math-expression
-                    // always update lastValueForDisplay
-                    // update rawRendererValue
-                    // if desired expression is different from math-expression obtained from current raw value
-                    // do not update immediate value
-
-                    instructions.push({
-                        setEssentialValue: "lastValueForDisplay",
-                        value: desiredStateVariableValues.rawRendererValue,
-                    });
-
-                    let currentMath = await calculateMathExpressionFromLatex(
-                        essentialValues.rawRendererValue,
-                    );
-
-                    // use deepCompare of trees rather than equalsViaSyntax
-                    // so even tiny numerical differences that within double precision are detected
-                    if (
-                        !deepCompare(
-                            desiredStateVariableValues.rawRendererValue.tree,
-                            currentMath.tree,
-                        )
-                    ) {
-                        let desiredValue = stripLatex(
-                            desiredStateVariableValues.rawRendererValue.toLatex(),
-                        );
-                        if (desiredValue === "\uff3f") {
-                            desiredValue = "";
-                        }
-                        instructions.push({
-                            setEssentialValue: "rawRendererValue",
-                            value: desiredValue,
-                        });
-                    }
-                }
-
-                return {
-                    success: true,
-                    instructions,
-                };
-            },
-        };
-
-        stateVariableDefinitions.componentType = {
-            returnDependencies: () => ({}),
-            definition: () => ({ setValue: { componentType: "math" } }),
-        };
-
-        stateVariableDefinitions.focused = {
-            forRenderer: true,
-            hasEssential: true,
-            defaultValue: false,
-            public: true,
-            description:
-                "Whether this part of the fraction input currently has keyboard focus.",
-            shadowingInstructions: {
-                createComponentOfType: "boolean",
-            },
-            ignoreFixed: true,
-            returnDependencies: () => ({}),
-            definition: () => ({
-                useEssentialOrDefaultValue: { focused: true },
-            }),
-            inverseDefinition({ desiredStateVariableValues }) {
-                return {
-                    success: true,
-                    instructions: [
-                        {
-                            setEssentialValue: "focused",
-                            value: Boolean(desiredStateVariableValues.focused),
-                        },
-                    ],
-                };
-            },
-        };
+        // valueForDisplay, text, and rawRendererValue
+        Object.assign(
+            stateVariableDefinitions,
+            returnMathComponentInputDisplayStateVariableDefinitions(),
+        );
 
         // Provide an accessible name/description for each part's textarea.
         stateVariableDefinitions.shortDescription = {
@@ -1350,116 +774,5 @@ export default class FractionComponentInput extends BaseComponent {
                 skipRendererUpdate,
             });
         }
-    }
-
-    async updateValue({
-        actionId,
-        sourceInformation = {},
-        skipRendererUpdate = false,
-    }) {
-        if (!(await this.stateValues.disabled)) {
-            let immediateValue = await this.stateValues.immediateValue;
-
-            if (
-                !deepCompare(
-                    (await this.stateValues.value).tree,
-                    immediateValue.tree,
-                )
-            ) {
-                let updateInstructions = [
-                    {
-                        updateType: "updateValue",
-                        componentIdx: this.componentIdx,
-                        stateVariable: "value",
-                        value: immediateValue,
-                    },
-                    // in case value ended up being a different value than requested
-                    // we set immediate value to whatever was the result
-                    // (hence the need to execute update first)
-                    {
-                        updateType: "executeUpdate",
-                    },
-                    {
-                        updateType: "updateValue",
-                        componentIdx: this.componentIdx,
-                        stateVariable: "immediateValue",
-                        valueOfStateVariable: "value",
-                    },
-                    {
-                        updateType: "unsetComponentNeedingUpdateValue",
-                    },
-                ];
-
-                if (immediateValue.tree !== "\uff3f") {
-                    updateInstructions.push({
-                        updateType: "updateValue",
-                        componentIdx: this.componentIdx,
-                        stateVariable: "rawRendererValue",
-                        valueOfStateVariable: "valueForDisplay",
-                    });
-                }
-
-                let event = {
-                    verb: "answered",
-                    object: {
-                        componentIdx: this.componentIdx,
-                        componentType: this.componentType,
-                    },
-                    result: {
-                        response: immediateValue,
-                        responseText: immediateValue.toString(),
-                    },
-                };
-
-                await this.coreFunctions.performUpdate({
-                    updateInstructions,
-                    actionId,
-                    sourceInformation,
-                    skipRendererUpdate: true,
-                    event,
-                });
-
-                return await this.coreFunctions.triggerChainedActions({
-                    componentIdx: this.componentIdx,
-                    actionId,
-                    sourceInformation,
-                    skipRendererUpdate,
-                });
-            } else {
-                // set raw renderer value to save it to the database,
-                // as it might not have been saved
-                // given that updateRawValue is transient
-                await this.coreFunctions.performUpdate({
-                    updateInstructions: [
-                        {
-                            updateType: "updateValue",
-                            componentIdx: this.componentIdx,
-                            stateVariable: "rawRendererValue",
-                            valueOfStateVariable: "rawRendererValue",
-                        },
-                    ],
-                    actionId,
-                    sourceInformation,
-                    skipRendererUpdate,
-                });
-            }
-        }
-    }
-
-    async focusChanged({ focused, actionId, sourceInformation }) {
-        return await this.coreFunctions.performUpdate({
-            updateInstructions: [
-                {
-                    updateType: "updateValue",
-                    componentIdx: this.componentIdx,
-                    stateVariable: "focused",
-                    value: focused,
-                },
-            ],
-            actionId,
-            sourceInformation,
-            overrideReadOnly: true,
-            doNotSave: true,
-        });
     }
 }
